Pull out query creation for paginated queries
