+++ app/routes/submissions/index.js
@@ -1,5 +1,6 @@
 import { service } from '@ember/service';
 import CheckSessionRoute from '../check-session-route';
+import { submissionsIndexQuery } from '../../util/paginated-query';
 
 export default class IndexRoute extends CheckSessionRoute {
   @service('current-user') currentUser;
@@ -12,37 +13,7 @@ export default class IndexRoute extends CheckSessionRoute {
   };
 
   async model(params) {
-    let query;
-
-    const user = this.currentUser.user;
-
-    if (user.isAdmin) {
-      query = {
-        filter: { submission: 'submissionStatus=out=cancelled' },
-        include: 'publication',
-      };
-    } else if (user.isSubmitter) {
-      const userMatch = `submitter.id==${user.id},preparers.id=in=${user.id}`;
-      query = {
-        filter: {
-          submission: `(${userMatch});submissionStatus=out=cancelled`,
-        },
-        sort: '-submittedDate',
-        include: 'publication',
-      };
-    }
-
-    const { page = 1, pageSize = 10 } = params;
-    query.page = {
-      number: page,
-      size: pageSize,
-      totals: true,
-    };
-
-    if (params.filter) {
-      query.filter.submission = `(${query.filter.submission});publication.title=ini=*${params.filter}*`;
-    }
-
+    const query = submissionsIndexQuery(params, this.currentUser.user);
     return this.store.query('submission', query).then((data) => ({
       submissions: data,
       meta: data.meta,
